Refactor for better folder structure
