+++ .eslintrc.json
@@ -170,6 +170,7 @@
             }
         ],
         "space-before-function-paren": "off",
+        "@typescript-eslint/no-var-requires": "warn",
         "@typescript-eslint/space-before-function-paren": [
             "error",
             "always"
