--- src/bin.ts
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-import './generator';
